requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.40.60
botocore==1.40.60
certifi==2026.1.4
click==8.3.1
fal_client==0.12.0
fastapi==0.128.0
gunicorn==24.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==26.0
Pillow==10.4.0
pydantic==2.12.5
pydantic_core==2.41.5
pytest==8.4.2
python-dotenv==1.2.1
python-multipart==0.0.9
s3transfer==0.14.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
websockets==16.0
